Split command palette support logic
