+++ supacode/Features/CommandPalette/Reducer/CommandPaletteFeature.swift
@@ -884,1092 +884,3 @@ private func makeClosePullRequestItem(
     ]
   }
 #endif
-
-private enum CommandPaletteItemID {
-  static let ghosttyPrefix = "ghostty."
-  static let globalCheckForUpdates = "global.check-for-updates"
-  static let globalOpenSettings = "global.open-settings"
-  static let globalOpenRepository = "global.open-repository"
-  static let globalNewWorktree = "global.new-worktree"
-  static let globalRefreshWorktrees = "global.refresh-worktrees"
-  static let globalJumpToLatestUnread = "global.jump-to-latest-unread"
-  static let globalViewArchivedWorktrees = "global.view-archived-worktrees"
-  static let globalInstallCLI = "global.install-cli"
-  static let globalToggleLeftSidebar = "global.toggle-left-sidebar"
-  static let globalToggleActiveAgentsPanel = "global.toggle-active-agents-panel"
-  static let globalToggleCanvas = "global.toggle-canvas"
-  static let globalExpandCanvasCard = "global.expand-canvas-card"
-  static let globalArrangeCanvasCards = "global.arrange-canvas-cards"
-  static let globalOrganizeCanvasCards = "global.organize-canvas-cards"
-  static let globalSelectAllCanvasCards = "global.select-all-canvas-cards"
-  static let globalToggleShelf = "global.toggle-shelf"
-  static let globalShowDiff = "global.show-diff"
-  static let globalRevealInFinder = "global.reveal-in-finder"
-  static let globalCopyPath = "global.copy-path"
-  static let globalRevealInSidebar = "global.reveal-in-sidebar"
-  static let globalRunScript = "global.run-script"
-  static let globalStopRunScript = "global.stop-run-script"
-  static let globalTogglePinWorktree = "global.toggle-pin-worktree"
-  static let globalRenameBranch = "global.rename-branch"
-  static let globalDeleteWorktree = "global.delete-worktree"
-
-  static func openRepositorySettings(_ repositoryID: Repository.ID) -> CommandPaletteItem.ID {
-    "repo.\(repositoryID).open-settings"
-  }
-
-  static func customCommand(_ commandID: String) -> CommandPaletteItem.ID {
-    "custom-command.\(commandID)"
-  }
-
-  static var globalIDs: [CommandPaletteItem.ID] {
-    [
-      globalCheckForUpdates,
-      globalOpenSettings,
-      globalOpenRepository,
-      globalNewWorktree,
-      globalRefreshWorktrees,
-      globalJumpToLatestUnread,
-      globalViewArchivedWorktrees,
-      globalInstallCLI,
-      globalToggleLeftSidebar,
-      globalToggleActiveAgentsPanel,
-      globalToggleCanvas,
-      globalExpandCanvasCard,
-      globalArrangeCanvasCards,
-      globalOrganizeCanvasCards,
-      globalSelectAllCanvasCards,
-      globalToggleShelf,
-      globalShowDiff,
-      globalRevealInFinder,
-      globalCopyPath,
-      globalRevealInSidebar,
-      globalRunScript,
-      globalStopRunScript,
-      globalTogglePinWorktree,
-      globalRenameBranch,
-      globalDeleteWorktree,
-    ]
-  }
-
-  static func worktreeSelect(_ worktreeID: Worktree.ID) -> CommandPaletteItem.ID {
-    "worktree.\(worktreeID).select"
-  }
-
-  static func changeFocusedTabIcon(_ worktreeID: Worktree.ID) -> CommandPaletteItem.ID {
-    "terminal.\(worktreeID).change-focused-tab-icon"
-  }
-
-  static func ghosttyCommand(_ command: GhosttyCommand) -> CommandPaletteItem.ID {
-    "\(ghosttyPrefix)\(command.action)|\(command.title)"
-  }
-
-  static func pullRequestIDs(repositoryID: Repository.ID) -> [CommandPaletteItem.ID] {
-    [
-      pullRequestOpen(repositoryID),
-      pullRequestReady(repositoryID),
-      pullRequestCopyFailingJobURL(repositoryID),
-      pullRequestCopyCiLogs(repositoryID),
-      pullRequestRerunFailedJobs(repositoryID),
-      pullRequestOpenFailingCheck(repositoryID),
-      pullRequestMerge(repositoryID),
-      pullRequestClose(repositoryID),
-    ]
-  }
-
-  static func pullRequestOpen(_ repositoryID: Repository.ID) -> CommandPaletteItem.ID {
-    "pr.\(repositoryID).open"
-  }
-
-  static func pullRequestReady(_ repositoryID: Repository.ID) -> CommandPaletteItem.ID {
-    "pr.\(repositoryID).ready"
-  }
-
-  static func pullRequestCopyFailingJobURL(_ repositoryID: Repository.ID) -> CommandPaletteItem.ID {
-    "pr.\(repositoryID).copy-failing-job-url"
-  }
-
-  static func pullRequestCopyCiLogs(_ repositoryID: Repository.ID) -> CommandPaletteItem.ID {
-    "pr.\(repositoryID).copy-ci-logs"
-  }
-
-  static func pullRequestRerunFailedJobs(_ repositoryID: Repository.ID) -> CommandPaletteItem.ID {
-    "pr.\(repositoryID).rerun-failed-jobs"
-  }
-
-  static func pullRequestOpenFailingCheck(_ repositoryID: Repository.ID) -> CommandPaletteItem.ID {
-    "pr.\(repositoryID).open-failing-check"
-  }
-
-  static func pullRequestMerge(_ repositoryID: Repository.ID) -> CommandPaletteItem.ID {
-    "pr.\(repositoryID).merge"
-  }
-
-  static func pullRequestClose(_ repositoryID: Repository.ID) -> CommandPaletteItem.ID {
-    "pr.\(repositoryID).close"
-  }
-}
-
-private func commandPaletteRecencyScore(
-  _ item: CommandPaletteItem,
-  recencyByID: [CommandPaletteItem.ID: TimeInterval],
-  now: Date
-) -> Double {
-  guard let lastActivated = recencyByID[item.id] else { return 0 }
-  let ageSeconds = max(0, now.timeIntervalSince1970 - lastActivated)
-  let ageDays = ageSeconds / 86_400
-  let cappedAgeDays = min(ageDays, 30)
-  return pow(0.5, cappedAgeDays / 7)
-}
-
-private func delegateAction(for kind: CommandPaletteItem.Kind) -> CommandPaletteFeature.Delegate {
-  if let appAction = appDelegateAction(for: kind) {
-    return appAction
-  }
-  switch kind {
-  case .worktreeSelect(let id):
-    return .selectWorktree(id)
-  case .deleteWorktree(let worktreeID, let repositoryID):
-    return .deleteWorktree(worktreeID, repositoryID)
-  case .ghosttyCommand(let action):
-    return .ghosttyCommand(action)
-  case .changeFocusedTabIcon(let worktreeID):
-    return .changeFocusedTabIcon(worktreeID)
-  case .togglePinWorktree(let worktreeID, let isCurrentlyPinned):
-    return .togglePinWorktree(worktreeID, isCurrentlyPinned: isCurrentlyPinned)
-  case .openRepositorySettings(let repositoryID):
-    return .openRepositorySettings(repositoryID)
-  case .runCustomCommand(let index, _, _):
-    return .runCustomCommand(index)
-  case .openPullRequest,
-    .openRepositoryOnCodeHost,
-    .markPullRequestReady,
-    .mergePullRequest,
-    .closePullRequest,
-    .copyFailingJobURL,
-    .copyCiFailureLogs,
-    .rerunFailedJobs,
-    .openFailingCheckDetails:
-    return pullRequestDelegateAction(for: kind)!
-  #if DEBUG
-    case .debugTestToast(let toast):
-      return .debugTestToast(toast)
-    case .debugSimulateUpdateFound:
-      return .debugSimulateUpdateFound
-    case .debugLightDockNotificationDot:
-      return .debugLightDockNotificationDot
-  #endif
-  case .checkForUpdates,
-    .openSettings,
-    .newWorktree,
-    .openRepository,
-    .viewArchivedWorktrees,
-    .refreshWorktrees,
-    .jumpToLatestUnread,
-    .installCLI,
-    .toggleLeftSidebar,
-    .toggleActiveAgentsPanel,
-    .toggleCanvas,
-    .expandCanvasCard,
-    .arrangeCanvasCards,
-    .organizeCanvasCards,
-    .selectAllCanvasCards,
-    .toggleShelf,
-    .showDiff,
-    .revealInFinder,
-    .copyPath,
-    .revealInSidebar,
-    .runScript,
-    .stopRunScript,
-    .renameBranch:
-    fatalError("appDelegateAction should handle app-level command palette actions")
-  }
-}
-
-private func appDelegateAction(for kind: CommandPaletteItem.Kind) -> CommandPaletteFeature.Delegate? {
-  if let delegate = navigationDelegateAction(for: kind) {
-    return delegate
-  }
-  if let delegate = viewDelegateAction(for: kind) {
-    return delegate
-  }
-  switch kind {
-  case .checkForUpdates:
-    return .checkForUpdates
-  case .openSettings:
-    return .openSettings
-  case .newWorktree:
-    return .newWorktree
-  case .openRepository:
-    return .openRepository
-  case .viewArchivedWorktrees:
-    return .viewArchivedWorktrees
-  case .refreshWorktrees:
-    return .refreshWorktrees
-  case .jumpToLatestUnread:
-    return .jumpToLatestUnread
-  case .installCLI:
-    return .installCLI
-  case .runScript:
-    return .runScript
-  case .stopRunScript:
-    return .stopRunScript
-  case .renameBranch:
-    return .renameBranch
-  default:
-    return nil
-  }
-}
-
-private func navigationDelegateAction(for kind: CommandPaletteItem.Kind) -> CommandPaletteFeature.Delegate? {
-  switch kind {
-  case .revealInFinder:
-    return .revealInFinder
-  case .copyPath:
-    return .copyPath
-  case .revealInSidebar:
-    return .revealInSidebar
-  default:
-    return nil
-  }
-}
-
-private func viewDelegateAction(for kind: CommandPaletteItem.Kind) -> CommandPaletteFeature.Delegate? {
-  switch kind {
-  case .toggleLeftSidebar:
-    return .toggleLeftSidebar
-  case .toggleActiveAgentsPanel:
-    return .toggleActiveAgentsPanel
-  case .toggleCanvas:
-    return .toggleCanvas
-  case .expandCanvasCard:
-    return .expandCanvasCard
-  case .arrangeCanvasCards:
-    return .arrangeCanvasCards
-  case .organizeCanvasCards:
-    return .organizeCanvasCards
-  case .selectAllCanvasCards:
-    return .selectAllCanvasCards
-  case .toggleShelf:
-    return .toggleShelf
-  case .showDiff:
-    return .showDiff
-  default:
-    return nil
-  }
-}
-
-private func pullRequestDelegateAction(
-  for kind: CommandPaletteItem.Kind
-) -> CommandPaletteFeature.Delegate? {
-  switch kind {
-  case .openPullRequest(let worktreeID),
-    .openRepositoryOnCodeHost(let worktreeID):
-    return .openPullRequest(worktreeID)
-  case .markPullRequestReady(let worktreeID):
-    return .markPullRequestReady(worktreeID)
-  case .mergePullRequest(let worktreeID):
-    return .mergePullRequest(worktreeID)
-  case .closePullRequest(let worktreeID):
-    return .closePullRequest(worktreeID)
-  case .copyFailingJobURL(let worktreeID):
-    return .copyFailingJobURL(worktreeID)
-  case .copyCiFailureLogs(let worktreeID):
-    return .copyCiFailureLogs(worktreeID)
-  case .rerunFailedJobs(let worktreeID):
-    return .rerunFailedJobs(worktreeID)
-  case .openFailingCheckDetails(let worktreeID):
-    return .openFailingCheckDetails(worktreeID)
-  case .worktreeSelect,
-    .checkForUpdates,
-    .openSettings,
-    .newWorktree,
-    .openRepository,
-    .viewArchivedWorktrees,
-    .refreshWorktrees,
-    .jumpToLatestUnread,
-    .installCLI,
-    .ghosttyCommand,
-    .changeFocusedTabIcon,
-    .toggleLeftSidebar,
-    .toggleActiveAgentsPanel,
-    .toggleCanvas,
-    .expandCanvasCard,
-    .arrangeCanvasCards,
-    .organizeCanvasCards,
-    .selectAllCanvasCards,
-    .toggleShelf,
-    .showDiff,
-    .revealInFinder,
-    .copyPath,
-    .revealInSidebar,
-    .runScript,
-    .stopRunScript,
-    .togglePinWorktree,
-    .renameBranch,
-    .deleteWorktree,
-    .openRepositorySettings,
-    .runCustomCommand:
-    return nil
-  #if DEBUG
-    case .debugTestToast, .debugSimulateUpdateFound, .debugLightDockNotificationDot:
-      return nil
-  #endif
-  }
-}
-
-/// Ghostty action keys that should be hidden from Prowl's command palette.
-///
-/// Includes:
-/// - actions Prowl already exposes natively (`check_for_updates`)
-/// - Ghostty actions intentionally unsupported by Prowl's architecture/platform
-private let filteredGhosttyActionKeys: Set<String> = [
-  "check_for_updates",
-  "new_window",
-  "close_all_windows",
-  "goto_window",
-  "toggle_tab_overview",
-  "toggle_window_decorations",
-  "inspector",
-  "show_gtk_inspector",
-  "show_on_screen_keyboard",
-]
-
-private func ghosttyCommandItems(_ commands: [GhosttyCommand]) -> [CommandPaletteItem] {
-  commands.compactMap { command in
-    guard !filteredGhosttyActionKeys.contains(command.actionKey) else { return nil }
-    return .ghosttyCommand(command)
-  }
-}
-
-extension CommandPaletteItem {
-  /// Build a top-level command backed by an `AppShortcuts` hotkey. Defaults to
-  /// `defaultSuggestion: true` (since these are the kinds of actions worth
-  /// listing when the palette opens with no query) and uses no subtitle (the
-  /// hotkey hint and title already do the work).
-  static func appShortcut(
-    id: String,
-    title: String,
-    category: Category,
-    kind: Kind,
-    keywords: [String] = [],
-    priorityTier: Int = defaultPriorityTier
-  ) -> CommandPaletteItem {
-    CommandPaletteItem(
-      id: id,
-      title: title,
-      subtitle: nil,
-      kind: kind,
-      category: category,
-      defaultSuggestion: true,
-      keywords: keywords,
-      priorityTier: priorityTier
-    )
-  }
-
-  /// Build an item wrapping a Ghostty command exposed via the runtime.
-  /// Lives in `.terminal`, ranks below regular suggestions (priority +100),
-  /// and is search-only (never surfaces on empty query).
-  static func ghosttyCommand(_ command: GhosttyCommand) -> CommandPaletteItem {
-    let subtitle = command.description.trimmingCharacters(in: .whitespacesAndNewlines)
-    return CommandPaletteItem(
-      id: CommandPaletteItemID.ghosttyCommand(command),
-      title: command.title,
-      subtitle: subtitle.isEmpty ? nil : subtitle,
-      kind: .ghosttyCommand(command.action),
-      category: .terminal,
-      defaultSuggestion: false,
-      priorityTier: CommandPaletteItem.defaultPriorityTier + 100
-    )
-  }
-}
-
-private func loadRecency(into state: inout CommandPaletteFeature.State) {
-  @Shared(.appStorage("commandPaletteItemRecency")) var recency: [String: Double] = [:]
-  state.recencyByItemID = recency
-}
-
-private func saveRecency(_ recencyByItemID: [CommandPaletteItem.ID: TimeInterval]) {
-  @Shared(.appStorage("commandPaletteItemRecency")) var recency: [String: Double] = [:]
-  $recency.withLock {
-    $0 = recencyByItemID
-  }
-}
-
-private struct CommandPaletteFuzzyScorer {
-  private struct PreparedQueryPiece {
-    let normalized: String
-    let normalizedLowercase: String
-    let expectContiguousMatch: Bool
-  }
-
-  private struct PreparedQuery {
-    let piece: PreparedQueryPiece
-    let values: [PreparedQueryPiece]?
-  }
-
-  private struct Match {
-    var start: Int
-    var end: Int
-  }
-
-  private struct ItemScore {
-    var score: Int
-    var labelMatch: [Match]?
-    var descriptionMatch: [Match]?
-  }
-
-  private struct ScoredItem {
-    let item: CommandPaletteItem
-    let score: ItemScore
-    let recencyScore: Double
-    let index: Int
-  }
-
-  private static let labelPrefixScoreThreshold = 1 << 17
-  private static let labelScoreThreshold = 1 << 16
-
-  private let query: PreparedQuery
-  private let allowNonContiguousMatches: Bool
-  private let recencyByID: [CommandPaletteItem.ID: TimeInterval]
-  private let now: Date
-
-  init(
-    query: String,
-    recencyByID: [CommandPaletteItem.ID: TimeInterval],
-    now: Date,
-    allowNonContiguousMatches: Bool = true
-  ) {
-    self.query = Self.prepareQuery(query)
-    self.allowNonContiguousMatches = allowNonContiguousMatches
-    self.recencyByID = recencyByID
-    self.now = now
-  }
-
-  func rankedItems(from items: [CommandPaletteItem]) -> [CommandPaletteItem] {
-    let scoredItems = items.enumerated().compactMap { index, item in
-      let score = scoreItem(item)
-      return score.score > 0
-        ? ScoredItem(
-          item: item,
-          score: score,
-          recencyScore: recencyScore(for: item),
-          index: index
-        )
-        : nil
-    }
-    let sorted = scoredItems.sorted { compare($0, $1) < 0 }
-    return sorted.map(\.item)
-  }
-
-  private func scoreItem(_ item: CommandPaletteItem) -> ItemScore {
-    guard !query.piece.normalized.isEmpty else {
-      return ItemScore(score: 0, labelMatch: nil, descriptionMatch: nil)
-    }
-
-    if let values = query.values, !values.isEmpty {
-      return scoreItemMultiple(
-        label: item.title,
-        description: item.subtitle,
-        keywords: item.keywords,
-        query: values
-      )
-    }
-
-    return scoreItemForPiece(
-      label: item.title,
-      description: item.subtitle,
-      keywords: item.keywords,
-      query: query.piece
-    )
-  }
-
-  private func scoreItemMultiple(
-    label: String,
-    description: String?,
-    keywords: [String],
-    query: [PreparedQueryPiece]
-  ) -> ItemScore {
-    var totalScore = 0
-    var totalLabelMatches: [Match] = []
-    var totalDescriptionMatches: [Match] = []
-
-    for piece in query {
-      let score = scoreItemForPiece(label: label, description: description, keywords: keywords, query: piece)
-      if score.score == 0 {
-        return ItemScore(score: 0, labelMatch: nil, descriptionMatch: nil)
-      }
-      totalScore += score.score
-      if let labelMatch = score.labelMatch {
-        totalLabelMatches.append(contentsOf: labelMatch)
-      }
-      if let descriptionMatch = score.descriptionMatch {
-        totalDescriptionMatches.append(contentsOf: descriptionMatch)
-      }
-    }
-
-    return ItemScore(
-      score: totalScore,
-      labelMatch: normalizeMatches(totalLabelMatches),
-      descriptionMatch: normalizeMatches(totalDescriptionMatches)
-    )
-  }
-
-  /// Score one query piece against the item's title (with description fallback)
-  /// and each keyword. Keywords act as alternative labels — when one outscores
-  /// the title path, we keep title-derived highlight positions so the UI never
-  /// paints offsets that don't exist in the visible string.
-  private func scoreItemForPiece(
-    label: String,
-    description: String?,
-    keywords: [String],
-    query: PreparedQueryPiece
-  ) -> ItemScore {
-    var best = scoreItemSingle(label: label, description: description, query: query)
-    for keyword in keywords {
-      let keywordScore = scoreItemSingle(label: keyword, description: nil, query: query)
-      if keywordScore.score > best.score {
-        best = ItemScore(
-          score: keywordScore.score,
-          labelMatch: best.labelMatch,
-          descriptionMatch: best.descriptionMatch
-        )
-      }
-    }
-    return best
-  }
-
-  private func scoreItemSingle(
-    label: String,
-    description: String?,
-    query: PreparedQueryPiece
-  ) -> ItemScore {
-    let (labelScore, labelPositions) = scoreFuzzy(
-      target: label,
-      query: query,
-      allowNonContiguousMatches: allowNonContiguousMatches && !query.expectContiguousMatch
-    )
-    if labelScore > 0 {
-      let labelPrefixMatch = matchesPrefix(query: query.normalizedLowercase, target: label)
-      let baseScore: Int
-      if let labelPrefixMatch {
-        let prefixLengthBoost = Int(
-          (Double(query.normalized.count) / Double(label.count) * 100).rounded()
-        )
-        baseScore = Self.labelPrefixScoreThreshold + prefixLengthBoost
-        return ItemScore(
-          score: baseScore + labelScore,
-          labelMatch: labelPrefixMatch,
-          descriptionMatch: nil
-        )
-      }
-      baseScore = Self.labelScoreThreshold
-      return ItemScore(
-        score: baseScore + labelScore,
-        labelMatch: createMatches(labelPositions),
-        descriptionMatch: nil
-      )
-    }
-
-    if let description {
-      let descriptionPrefixLength = description.count
-      let descriptionAndLabel = description + label
-      let (labelDescriptionScore, labelDescriptionPositions) = scoreFuzzy(
-        target: descriptionAndLabel,
-        query: query,
-        allowNonContiguousMatches: allowNonContiguousMatches && !query.expectContiguousMatch
-      )
-      if labelDescriptionScore > 0 {
-        let labelDescriptionMatches = createMatches(labelDescriptionPositions)
-        var labelMatch: [Match] = []
-        var descriptionMatch: [Match] = []
-
-        for match in labelDescriptionMatches {
-          if match.start < descriptionPrefixLength && match.end > descriptionPrefixLength {
-            labelMatch.append(Match(start: 0, end: match.end - descriptionPrefixLength))
-            descriptionMatch.append(Match(start: match.start, end: descriptionPrefixLength))
-          } else if match.start >= descriptionPrefixLength {
-            labelMatch.append(
-              Match(
-                start: match.start - descriptionPrefixLength,
-                end: match.end - descriptionPrefixLength
-              )
-            )
-          } else {
-            descriptionMatch.append(match)
-          }
-        }
-
-        return ItemScore(
-          score: labelDescriptionScore,
-          labelMatch: labelMatch,
-          descriptionMatch: descriptionMatch
-        )
-      }
-    }
-
-    return ItemScore(score: 0, labelMatch: nil, descriptionMatch: nil)
-  }
-
-  private func compare(_ itemA: ScoredItem, _ itemB: ScoredItem) -> Int {
-    let scoreA = itemA.score.score
-    let scoreB = itemB.score.score
-
-    if scoreA > Self.labelScoreThreshold || scoreB > Self.labelScoreThreshold {
-      if scoreA != scoreB {
-        return scoreA > scoreB ? -1 : 1
-      }
-      if scoreA < Self.labelPrefixScoreThreshold && scoreB < Self.labelPrefixScoreThreshold {
-        let comparedByMatchLength = compareByMatchLength(itemA.score.labelMatch, itemB.score.labelMatch)
-        if comparedByMatchLength != 0 {
-          return comparedByMatchLength
-        }
-      }
-      let labelA = itemA.item.title
-      let labelB = itemB.item.title
-      if labelA.count != labelB.count {
-        return labelA.count - labelB.count
-      }
-    }
-
-    if scoreA != scoreB {
-      return scoreA > scoreB ? -1 : 1
-    }
-
-    let itemAHasLabelMatches = !(itemA.score.labelMatch?.isEmpty ?? true)
-    let itemBHasLabelMatches = !(itemB.score.labelMatch?.isEmpty ?? true)
-    if itemAHasLabelMatches && !itemBHasLabelMatches {
-      return -1
-    }
-    if itemBHasLabelMatches && !itemAHasLabelMatches {
-      return 1
-    }
-
-    if let itemAMatchDistance = matchDistance(itemA),
-      let itemBMatchDistance = matchDistance(itemB),
-      itemAMatchDistance != itemBMatchDistance
-    {
-      return itemBMatchDistance > itemAMatchDistance ? -1 : 1
-    }
-
-    if itemA.item.priorityTier != itemB.item.priorityTier {
-      return itemA.item.priorityTier < itemB.item.priorityTier ? -1 : 1
-    }
-
-    if itemA.recencyScore != itemB.recencyScore {
-      return itemA.recencyScore > itemB.recencyScore ? -1 : 1
-    }
-
-    let fallback = fallbackCompare(itemA.item, itemB.item)
-    if fallback != 0 {
-      return fallback
-    }
-
-    return itemA.index - itemB.index
-  }
-
-  private func matchDistance(_ item: ScoredItem) -> Int? {
-    var matchStart = -1
-    var matchEnd = -1
-
-    if let descriptionMatch = item.score.descriptionMatch, !descriptionMatch.isEmpty {
-      matchStart = descriptionMatch[0].start
-    } else if let labelMatch = item.score.labelMatch, !labelMatch.isEmpty {
-      matchStart = labelMatch[0].start
-    }
-
-    if let labelMatch = item.score.labelMatch, !labelMatch.isEmpty {
-      matchEnd = labelMatch[labelMatch.count - 1].end
-      if let descriptionMatch = item.score.descriptionMatch,
-        !descriptionMatch.isEmpty,
-        let description = item.item.subtitle
-      {
-        matchEnd += description.count
-      }
-    } else if let descriptionMatch = item.score.descriptionMatch, !descriptionMatch.isEmpty {
-      matchEnd = descriptionMatch[descriptionMatch.count - 1].end
-    }
-
-    guard matchStart != -1 else { return nil }
-    return matchEnd - matchStart
-  }
-
-  private func compareByMatchLength(_ matchesA: [Match]?, _ matchesB: [Match]?) -> Int {
-    guard let matchesA, let matchesB else { return 0 }
-    if matchesA.isEmpty && matchesB.isEmpty {
-      return 0
-    }
-    if matchesB.isEmpty {
-      return -1
-    }
-    if matchesA.isEmpty {
-      return 1
-    }
-
-    let matchLengthA = matchesA[matchesA.count - 1].end - matchesA[0].start
-    let matchLengthB = matchesB[matchesB.count - 1].end - matchesB[0].start
-
-    if matchLengthA == matchLengthB {
-      return 0
-    }
-    return matchLengthB < matchLengthA ? 1 : -1
-  }
-
-  private func fallbackCompare(_ itemA: CommandPaletteItem, _ itemB: CommandPaletteItem) -> Int {
-    let labelA = itemA.title
-    let labelB = itemB.title
-    let descriptionA = itemA.subtitle
-    let descriptionB = itemB.subtitle
-
-    let labelDescriptionALength = labelA.count + (descriptionA?.count ?? 0)
-    let labelDescriptionBLength = labelB.count + (descriptionB?.count ?? 0)
-
-    if labelDescriptionALength != labelDescriptionBLength {
-      return labelDescriptionALength - labelDescriptionBLength
-    }
-
-    if labelA != labelB {
-      return compareStrings(labelA, labelB)
-    }
-
-    if let descriptionA, let descriptionB, descriptionA != descriptionB {
-      return compareStrings(descriptionA, descriptionB)
-    }
-
-    return 0
-  }
-
-  private func compareStrings(_ stringA: String, _ stringB: String) -> Int {
-    switch stringA.localizedStandardCompare(stringB) {
-    case .orderedAscending:
-      return -1
-    case .orderedDescending:
-      return 1
-    case .orderedSame:
-      return 0
-    }
-  }
-
-  private func recencyScore(for item: CommandPaletteItem) -> Double {
-    commandPaletteRecencyScore(item, recencyByID: recencyByID, now: now)
-  }
-
-  private func scoreFuzzy(
-    target: String,
-    query: PreparedQueryPiece,
-    allowNonContiguousMatches: Bool
-  ) -> (Int, [Int]) {
-    if target.isEmpty || query.normalized.isEmpty {
-      return (0, [])
-    }
-
-    let targetChars = Array(target)
-    let queryChars = Array(query.normalized)
-
-    if targetChars.count < queryChars.count {
-      return (0, [])
-    }
-
-    let targetLower = Array(target.lowercased())
-    let queryLower = Array(query.normalizedLowercase)
-
-    return doScoreFuzzy(
-      query: queryChars,
-      queryLower: queryLower,
-      target: targetChars,
-      targetLower: targetLower,
-      allowNonContiguousMatches: allowNonContiguousMatches
-    )
-  }
-
-  private func doScoreFuzzy(
-    query: [Character],
-    queryLower: [Character],
-    target: [Character],
-    targetLower: [Character],
-    allowNonContiguousMatches: Bool
-  ) -> (Int, [Int]) {
-    let queryLength = query.count
-    let targetLength = target.count
-    let scores = Array(repeating: 0, count: queryLength * targetLength)
-    var mutableScores = scores
-    let matches = Array(repeating: 0, count: queryLength * targetLength)
-    var mutableMatches = matches
-
-    for queryIndex in 0..<queryLength {
-      let queryIndexOffset = queryIndex * targetLength
-      let queryIndexPreviousOffset = queryIndexOffset - targetLength
-      let queryIndexGtNull = queryIndex > 0
-
-      let queryCharAtIndex = query[queryIndex]
-      let queryLowerCharAtIndex = queryLower[queryIndex]
-
-      for targetIndex in 0..<targetLength {
-        let targetIndexGtNull = targetIndex > 0
-
-        let currentIndex = queryIndexOffset + targetIndex
-        let leftIndex = currentIndex - 1
-        let diagIndex = queryIndexPreviousOffset + targetIndex - 1
-
-        let leftScore = targetIndexGtNull ? mutableScores[leftIndex] : 0
-        let diagScore = queryIndexGtNull && targetIndexGtNull ? mutableScores[diagIndex] : 0
-
-        let matchesSequenceLength =
-          queryIndexGtNull && targetIndexGtNull ? mutableMatches[diagIndex] : 0
-
-        let score: Int
-        let scoreContext = CharScoreContext(
-          queryChar: queryCharAtIndex,
-          queryLowerChar: queryLowerCharAtIndex,
-          target: target,
-          targetLower: targetLower,
-          targetIndex: targetIndex,
-          matchesSequenceLength: matchesSequenceLength
-        )
-        if diagScore != 0 && queryIndexGtNull {
-          score = computeCharScore(scoreContext)
-        } else if queryIndexGtNull {
-          score = 0
-        } else {
-          score = computeCharScore(scoreContext)
-        }
-
-        let isValidScore = score > 0 && diagScore + score >= leftScore
-
-        if isValidScore
-          && (allowNonContiguousMatches || queryIndexGtNull
-            || startsWith(
-              targetLower,
-              queryLower,
-              at: targetIndex
-            ))
-        {
-          mutableMatches[currentIndex] = matchesSequenceLength + 1
-          mutableScores[currentIndex] = diagScore + score
-        } else {
-          mutableMatches[currentIndex] = 0
-          mutableScores[currentIndex] = leftScore
-        }
-      }
-    }
-
-    var positions: [Int] = []
-    var queryIndex = queryLength - 1
-    var targetIndex = targetLength - 1
-    while queryIndex >= 0 && targetIndex >= 0 {
-      let currentIndex = queryIndex * targetLength + targetIndex
-      let match = mutableMatches[currentIndex]
-      if match == 0 {
-        targetIndex -= 1
-      } else {
-        positions.append(targetIndex)
-        queryIndex -= 1
-        targetIndex -= 1
-      }
-    }
-
-    positions.reverse()
-    let finalScore = mutableScores[queryLength * targetLength - 1]
-    return (finalScore, positions)
-  }
-
-  private struct CharScoreContext {
-    let queryChar: Character
-    let queryLowerChar: Character
-    let target: [Character]
-    let targetLower: [Character]
-    let targetIndex: Int
-    let matchesSequenceLength: Int
-  }
-
-  private func computeCharScore(_ context: CharScoreContext) -> Int {
-    if !considerAsEqual(context.queryLowerChar, context.targetLower[context.targetIndex]) {
-      return 0
-    }
-
-    var score = 1
-
-    if context.matchesSequenceLength > 0 {
-      score += (min(context.matchesSequenceLength, 3) * 6)
-      score += max(0, context.matchesSequenceLength - 3) * 3
-    }
-
-    if context.queryChar == context.target[context.targetIndex] {
-      score += 1
-    }
-
-    if context.targetIndex == 0 {
-      score += 8
-    } else {
-      let separatorBonus = scoreSeparatorAtPos(context.target[context.targetIndex - 1])
-      if separatorBonus > 0 {
-        score += separatorBonus
-      } else if isUpper(context.target[context.targetIndex]) && context.matchesSequenceLength == 0 {
-        score += 2
-      }
-    }
-
-    return score
-  }
-
-  private func considerAsEqual(_ lhs: Character, _ rhs: Character) -> Bool {
-    if lhs == rhs {
-      return true
-    }
-    if lhs == "/" || lhs == "\\" {
-      return rhs == "/" || rhs == "\\"
-    }
-    return false
-  }
-
-  private func scoreSeparatorAtPos(_ char: Character) -> Int {
-    switch char {
-    case "/", "\\":
-      return 5
-    case "_", "-", ".", " ", "'", "\"", ":":
-      return 4
-    default:
-      return 0
-    }
-  }
-
-  private func isUpper(_ char: Character) -> Bool {
-    guard let scalar = String(char).unicodeScalars.first else { return false }
-    return scalar.properties.isUppercase
-  }
-
-  private func startsWith(
-    _ target: [Character],
-    _ query: [Character],
-    at index: Int
-  ) -> Bool {
-    guard index + query.count <= target.count else { return false }
-    for queryIndex in 0..<query.count where target[index + queryIndex] != query[queryIndex] {
-      return false
-    }
-    return true
-  }
-
-  private func createMatches(_ offsets: [Int]) -> [Match] {
-    var matches: [Match] = []
-    var lastMatch: Match?
-
-    for position in offsets {
-      if var lastMatch, lastMatch.end == position {
-        lastMatch.end += 1
-        matches[matches.count - 1] = lastMatch
-      } else {
-        let match = Match(start: position, end: position + 1)
-        matches.append(match)
-        lastMatch = match
-      }
-    }
-
-    return matches
-  }
-
-  private func normalizeMatches(_ matches: [Match]) -> [Match]? {
-    guard !matches.isEmpty else { return nil }
-
-    let sortedMatches = matches.sorted { $0.start < $1.start }
-    var normalizedMatches: [Match] = []
-    var currentMatch: Match?
-
-    for match in sortedMatches {
-      if let existing = currentMatch, matchOverlaps(existing, match) {
-        let merged = Match(
-          start: min(existing.start, match.start),
-          end: max(existing.end, match.end)
-        )
-        currentMatch = merged
-        normalizedMatches[normalizedMatches.count - 1] = merged
-      } else {
-        currentMatch = match
-        normalizedMatches.append(match)
-      }
-    }
-
-    return normalizedMatches
-  }
-
-  private func matchOverlaps(_ matchA: Match, _ matchB: Match) -> Bool {
-    if matchA.end < matchB.start {
-      return false
-    }
-    if matchB.end < matchA.start {
-      return false
-    }
-    return true
-  }
-
-  private func matchesPrefix(query: String, target: String) -> [Match]? {
-    let targetLower = target.lowercased()
-    guard targetLower.hasPrefix(query) else { return nil }
-    return [Match(start: 0, end: query.count)]
-  }
-
-  private static func prepareQuery(_ original: String) -> PreparedQuery {
-    let expectContiguousMatch = queryExpectsExactMatch(original)
-    let normalized = normalizeQuery(original)
-    let piece = PreparedQueryPiece(
-      normalized: normalized.normalized,
-      normalizedLowercase: normalized.normalizedLowercase,
-      expectContiguousMatch: expectContiguousMatch
-    )
-
-    let splitPieces = original.split(separator: " ")
-    var values: [PreparedQueryPiece] = []
-    if splitPieces.count > 1 {
-      for pieceValue in splitPieces {
-        let value = String(pieceValue)
-        let expectExactMatchPiece = queryExpectsExactMatch(value)
-        let normalizedPiece = normalizeQuery(value)
-        if normalizedPiece.normalized.isEmpty {
-          continue
-        }
-        values.append(
-          PreparedQueryPiece(
-            normalized: normalizedPiece.normalized,
-            normalizedLowercase: normalizedPiece.normalizedLowercase,
-            expectContiguousMatch: expectExactMatchPiece
-          )
-        )
-      }
-    }
-
-    return PreparedQuery(
-      piece: piece,
-      values: values.isEmpty ? nil : values
-    )
-  }
-
-  private static func normalizeQuery(_ original: String) -> (normalized: String, normalizedLowercase: String) {
-    var pathNormalized = String()
-    pathNormalized.reserveCapacity(original.count)
-    for char in original {
-      if char == "\\" {
-        pathNormalized.append("/")
-      } else {
-        pathNormalized.append(char)
-      }
-    }
-
-    var normalized = String()
-    normalized.reserveCapacity(pathNormalized.count)
-    for char in pathNormalized {
-      if char == "*" || char == "…" || char == "\"" || char.isWhitespace {
-        continue
-      }
-      normalized.append(char)
-    }
-
-    if normalized.count > 1, normalized.hasSuffix("#") {
-      normalized.removeLast()
-    }
-
-    return (normalized, normalized.lowercased())
-  }
-
-  private static func queryExpectsExactMatch(_ query: String) -> Bool {
-    query.hasPrefix("\"") && query.hasSuffix("\"")
-  }
-}
